Using uint32_t instead of uint8_t type for bitfields inside struct definitions. The usage of uint8_t type birfields was preventing the code to compile.

+++ isotp_defines.h
@@ -63,33 +63,33 @@ typedef enum {
 /* can fram defination */
 #if defined(ISOTP_BYTE_ORDER_LITTLE_ENDIAN)
 typedef struct {
-    uint8_t reserve_1:4;
-    uint8_t type:4;
+    uint32_t reserve_1:4;
+    uint32_t type:4;
     uint8_t reserve_2[7];
 } IsoTpPciType;
 
 typedef struct {
-    uint8_t SF_DL:4;
-    uint8_t type:4;
+    uint32_t SF_DL:4;
+    uint32_t type:4;
     uint8_t data[7];
 } IsoTpSingleFrame;
 
 typedef struct {
-    uint8_t FF_DL_high:4;
-    uint8_t type:4;
+    uint32_t FF_DL_high:4;
+    uint32_t type:4;
     uint8_t FF_DL_low;
     uint8_t data[6];
 } IsoTpFirstFrame;
 
 typedef struct {
-    uint8_t SN:4;
-    uint8_t type:4;
+    uint32_t SN:4;
+    uint32_t type:4;
     uint8_t data[7];
 } IsoTpConsecutiveFrame;
 
 typedef struct {
-    uint8_t FS:4;
-    uint8_t type:4;
+    uint32_t FS:4;
+    uint32_t type:4;
     uint8_t BS;
     uint8_t STmin;
     uint8_t reserve[5];
@@ -98,8 +98,8 @@ typedef struct {
 #else
 
 typedef struct {
-    uint8_t type:4;
-    uint8_t reserve_1:4;
+    uint32_t type:4;
+    uint32_t reserve_1:4;
     uint8_t reserve_2[7];
 } IsoTpPciType;
 
@@ -114,8 +114,8 @@ typedef struct {
 * +-------------+-----------+-----+
 */
 typedef struct {
-    uint8_t type:4;
-    uint8_t SF_DL:4;
+    uint32_t type:4;
+    uint32_t SF_DL:4;
     uint8_t data[7];
 } IsoTpSingleFrame;
 
@@ -130,8 +130,8 @@ typedef struct {
 * +-------------+-----------+-----------------------+-----+
 */
 typedef struct {
-    uint8_t type:4;
-    uint8_t FF_DL_high:4;
+    uint32_t type:4;
+    uint32_t FF_DL_high:4;
     uint8_t FF_DL_low;
     uint8_t data[6];
 } IsoTpFirstFrame;
@@ -147,8 +147,8 @@ typedef struct {
 * +-------------+-----------+-----+
 */
 typedef struct {
-    uint8_t type:4;
-    uint8_t SN:4;
+    uint32_t type:4;
+    uint32_t SN:4;
     uint8_t data[7];
 } IsoTpConsecutiveFrame;
 
@@ -163,8 +163,8 @@ typedef struct {
 * +-------------+-----------+-----------------------+-----------------------+-----+
 */
 typedef struct {
-    uint8_t type:4;
-    uint8_t FS:4;
+    uint32_t type:4;
+    uint32_t FS:4;
     uint8_t BS;
     uint8_t STmin;
     uint8_t reserve[5];
